organize simulation examples into modules
