+++ src/SocialSimulations.jl
@@ -7,5 +7,8 @@ module SocialSimulations
 
     include("simulations/AbstractSocialSimulation.jl")
     include("simulations/SocialSimulation.jl")
+    
+    include("simulations/Dummy.jl")
+    include("simulations/LoneParentsModel.jl")
 
 end # Socialimulations
