+++ src/simulations/LoneParentsModel.jl
@@ -1,0 +1,41 @@
+"""
+    just a list of [stepping] functions to operate a Multi ABM 
+    design of function arguments and return values to be specified later
+""" 
+
+module LoneParentsModel
+
+    export createPopulation
+
+    import SocialABMs: SocialABM 
+    import SocialAgents: Person
+
+    "Create an empty population initially with no agents"
+    function createPopulation() 
+        population = SocialABM{Person}()
+    end
+
+    function doDeaths() end 
+    function doAdoptions() end 
+    function oneTransition_UCN() end
+    function computeIncome() end 
+    function updateWealthInd() end
+    function jobMarket() end
+    function computeAggregateSchedule() end 
+    function startCareAllocation() end 
+    function allocateWeeklyCare() end 
+    function updateUnmetCare() end 
+    function updateUnmetCareNeed() end 
+    function doAgeTransition() end 
+    function doSocialTransitions() end 
+    function houseOwnership() end 
+    function computeBenefits() end 
+    function doBirths() end 
+    function doDevorces() end 
+    function doMarriages() end 
+    function doMovingAround() end
+    function selfPyramidUpdate() end
+    function healthcareCost() end
+    function doStats() end  
+
+end # Module LoneParentsModel
